+++ src/main/java/com/buuz135/functionalstorage/data/FunctionalStorageLangProvider.java
@@ -131,6 +131,7 @@ public class FunctionalStorageLangProvider extends LanguageProvider {
         this.add("gui.functionalstorage.fluid_drawer_hint.empty", "%s: Empty container into drawer");
         this.add("gui.functionalstorage.fluid_drawer_hint.fill", "%s: Fill container from drawer");
         this.add("gui.functionalstorage.open_gui", "Right Click while Crouching to open the GUI");
+        this.add("gui.functionalstorage.priority", "Priority");
         this.add("gui.functionalstorage.storage", "Storage");
         this.add("gui.functionalstorage.utility", "Utility");
         this.add("gui.functionalstorage.storage_range", "Range");